+++ codes/codes_test.go
@@ -25,11 +25,10 @@ import (
 
 	cpb "google.golang.org/genproto/googleapis/rpc/code"
 	"google.golang.org/grpc/internal/grpctest"
-	"google.golang.org/grpc/internal/grpctest/tlogger"
 )
 
 type s struct {
-	tlogger.Tester
+	grpctest.Tester
 }
 
 func Test(t *testing.T) {
