+++ credentials/internal/syscallconn_test.go
@@ -27,11 +27,10 @@ import (
 
 	"google.golang.org/grpc/credentials/internal"
 	"google.golang.org/grpc/internal/grpctest"
-	"google.golang.org/grpc/internal/grpctest/tlogger"
 )
 
 type s struct {
-	tlogger.Tester
+	grpctest.Tester
 }
 
 func Test(t *testing.T) {
