+++ internal/buffer/unbounded_test.go
@@ -24,7 +24,6 @@ import (
 	"testing"
 
 	"google.golang.org/grpc/internal/grpctest"
-	"google.golang.org/grpc/internal/grpctest/tlogger"
 )
 
 const (
@@ -33,7 +32,7 @@ const (
 )
 
 type s struct {
-	tlogger.Tester
+	grpctest.Tester
 }
 
 func Test(t *testing.T) {
