+++ internal/grpctest/grpctest.go
@@ -22,8 +22,44 @@ package grpctest
 import (
 	"reflect"
 	"strings"
+	"sync/atomic"
 	"testing"
+
+	"google.golang.org/grpc/internal/leakcheck"
 )
+
+var lcFailed uint32
+
+type errorer struct {
+	t *testing.T
+}
+
+func (e errorer) Errorf(format string, args ...interface{}) {
+	atomic.StoreUint32(&lcFailed, 1)
+	e.t.Errorf(format, args...)
+}
+
+// Tester is an implementation of the x interface parameter to
+// grpctest.RunSubTests with default Setup and Teardown behavior. Setup updates
+// the tlogger and Teardown performs a leak check. Embed in a struct with tests
+// defined to use.
+type Tester struct{}
+
+// Setup updates the tlogger.
+func (Tester) Setup(t *testing.T) {
+	Update(t)
+}
+
+// Teardown performs a leak check.
+func (Tester) Teardown(t *testing.T) {
+	if atomic.LoadUint32(&lcFailed) == 1 {
+		return
+	}
+	leakcheck.Check(errorer{t: t})
+	if atomic.LoadUint32(&lcFailed) == 1 {
+		t.Log("Leak check disabled for future tests")
+	}
+}
 
 func getTestFunc(t *testing.T, xv reflect.Value, name string) func(*testing.T) {
 	if m := xv.MethodByName(name); m.IsValid() {
